requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.4
Flask-Testing==0.8.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
packaging==24.1
pluggy==1.5.0
PyMySQL==1.0.3
pytest==8.3.3
python-dotenv==1.0.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
#